move fetch utils to own file
